server/requirements.txt
colorama==0.4.6
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pytest==8.3.5
websockets==15.0.1